#32: Added documentation, updated jsdoc
