+++ lib/directive.js
@@ -169,7 +169,6 @@ _.extend(Slingshot.Directive.prototype, {
   * @param {Object} context
   * @param {FileInfo} file
   * @param {Object} [meta]
-  * @param {Function} [authorize]
   *
   * @returns {Boolean}
   */
