+++ lib/upload.js
@@ -60,6 +60,11 @@ Slingshot.Upload = function (directive, metaData) {
       return loaded.get();
     },
     
+   /**
+    * @param {File} file
+    * @returns {Boolean|Error} Returns to on success, Error on failure.
+    */
+    
     validate: function(file) {
       var context = {
         userId: Meteor.userId()
